Create cvfs.c
This project is derived from concept of Linux Operating System and based on how Files are managed in Operating System .A File is common storage unit in computer. File is an unformatted uniform stream of bytes. Bringing file to the RAM is done by operating system, this task is called as File Management. File Management is responsible for file handling system calls like create, open, close, read, write, lseek etc. In this project, I have implemented my own system calls for read, write, create etc. In this file System, no permanent file will be created on hard-disk; rather it will be created on RAM, and will be disappear as the program terminates. But we can read, write, seek file data from all the files using function I have implemented. This file system have all the section of hard-disk like super block, DILB and data block. Also it have UFDT Area, File table and IIT Table. There are 4 structure 1: ufdt 2: Filetable 3: inode  4: Superblock. I have created the structure of ufdt in which there is a pointer holding address of file table. I have made the array of this ufdt structure. pointer of filetable points to inode and inode is a Singly Linked List which has a member known as buffer in which our data is preserved. There are total 50 inode in this file system. All the information of files in computer is stored in inode. This inode is a self referential data-structure. There are total 50 inodes in this file system hence we can create total 50 files. Buffer is char array of size 1024 Bytes in which our data will be stored by dynamically allocation. We read, write data in Buffer and we can write total 1024bytes of data in one file. To read a file give file mode 1 and to write give mode 2. And if we want to both read and write file give mode 3(1+2).